+++ CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [8.9.0]
 
 ### Added
-- Added the `exh.tasks.functions` methods `find`, `getByName`, `update`, `enable`, `disable` and `remove`
+- Added the `exh.tasks.functions` methods `find`, `findByName`, `update`, `enable`, `disable` and `remove`
 - Added a `exh.users.findByEmail` method to find a user by their email address.
 - Completed the `Task` type for changes made in TaskService 1.3.0 a while ago:
   - Added the `retriedByTaskId`, `retryForTaskIds`, and `error` properties.
@@ -19,6 +19,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Deprecated
 - `exh.tasks.schedules.delete` is deprecated in favor of `exh.tasks.schedules.remove`
 - `exh.auth.oidc.providers.delete` is deprecated in favor of `exh.auth.oidc.providers.remove`
+- `exh.notificationsV2.userSettings.getById` is deprecated in favor of `exh.notificationsV2.userSettings.findById`
+- `exh.localizations.getByKeys` is deprecated in favor of `exh.localizations.findByKeys`
 
 ### Fixed
 - Corrected the types for the `findFirst` (based) methods to indicate they may return `undefined` when no results are found.